Merge 21.0 into 21.1-dev

See merge request atlas/athena!9100

Former-commit-id: a6135930d627cc690cf3a1ad20f3fdad775c59d5

+++ Tools/PyJobTransforms/python/trfMPTools.py
@@ -61,42 +61,49 @@ def detectAthenaMPProcs(argdict = {}):
 #  @param athenaMPworkers Number of AthenaMP workers
 #  @param skipFileChecks Switches off checks on output files
 #  @return @c None; side effect is the update of the @c dataDictionary
-def athenaMPOutputHandler(athenaMPFileReport, athenaMPWorkerTopDir, dataDictionary, athenaMPworkers, skipFileChecks):
+def athenaMPOutputHandler(athenaMPFileReport, athenaMPWorkerTopDir, dataDictionary, athenaMPworkers, skipFileChecks = False, argdict = {}):
     msg.debug("MP output handler called for report {0} and workers in {1}, data types {2}".format(athenaMPFileReport, athenaMPWorkerTopDir, dataDictionary.keys()))
     outputHasBeenHandled = dict([ (dataType, False) for dataType in dataDictionary.keys() if dataDictionary[dataType] ])
 
-    # First, see what AthenaMP told us
-    mpOutputs = ElementTree.ElementTree()
-    try:
-        mpOutputs.parse(athenaMPFileReport)
-    except IOError:
-        raise trfExceptions.TransformExecutionException(trfExit.nameToCode("TRF_OUTPUT_FILE_ERROR"), "Missing AthenaMP outputs file {0} (probably athena crashed)".format(athenaMPFileReport))
-    for filesElement in mpOutputs.getroot().getiterator(tag='Files'):
-        msg.debug('Examining element {0} with attributes {1}'.format(filesElement, filesElement.attrib))
-        originalArg = None 
-        startName = filesElement.attrib['OriginalName']
-        for dataType, fileArg in dataDictionary.iteritems():
-            if fileArg.value[0] == startName:
-                originalArg = fileArg
-                outputHasBeenHandled[dataType] = True
-                break
-        if originalArg is None:
-            msg.warning('Found AthenaMP output with name {0}, but no matching transform argument'.format(startName))
-            continue
+    # if sharedWriter mode is active ignore athenaMPFileReport
+    sharedWriter=False
+    if 'sharedWriter' in argdict and argdict['sharedWriter'].value:
+        sharedWriter=True
+        skipFileChecks=True
+
+    if not sharedWriter:
+        # First, see what AthenaMP told us
+        mpOutputs = ElementTree.ElementTree()
+        try:
+            mpOutputs.parse(athenaMPFileReport)
+        except IOError:
+            raise trfExceptions.TransformExecutionException(trfExit.nameToCode("TRF_OUTPUT_FILE_ERROR"), "Missing AthenaMP outputs file {0} (probably athena crashed)".format(athenaMPFileReport))
+        for filesElement in mpOutputs.getroot().getiterator(tag='Files'):
+            msg.debug('Examining element {0} with attributes {1}'.format(filesElement, filesElement.attrib))
+            originalArg = None 
+            startName = filesElement.attrib['OriginalName']
+            for dataType, fileArg in dataDictionary.iteritems():
+                if fileArg.value[0] == startName:
+                    originalArg = fileArg
+                    outputHasBeenHandled[dataType] = True
+                    break
+            if originalArg is None:
+                msg.warning('Found AthenaMP output with name {0}, but no matching transform argument'.format(startName))
+                continue
         
-        msg.debug('Found matching argument {0}'.format(originalArg))
-        fileNameList = []
-        for fileElement in filesElement.getiterator(tag='File'):
-            msg.debug('Examining element {0} with attributes {1}'.format(fileElement, fileElement.attrib))
-            fileNameList.append(path.relpath(fileElement.attrib['name']))
-            
-        athenaMPoutputsLinkAndUpdate(fileNameList, fileArg) 
-        
+            msg.debug('Found matching argument {0}'.format(originalArg))
+            fileNameList = []
+            for fileElement in filesElement.getiterator(tag='File'):
+                msg.debug('Examining element {0} with attributes {1}'.format(fileElement, fileElement.attrib))
+                fileNameList.append(path.relpath(fileElement.attrib['name']))
+
+            athenaMPoutputsLinkAndUpdate(fileNameList, fileArg)
+
     # Now look for additional outputs that have not yet been handled
     if len([ dataType for dataType in outputHasBeenHandled if outputHasBeenHandled[dataType] is False]):
         # OK, we have something we need to search for; cache the dirwalk here
         MPdirWalk = [ dirEntry for dirEntry in os.walk(athenaMPWorkerTopDir) ]
-        
+
         for dataType, fileArg in dataDictionary.iteritems():
             if outputHasBeenHandled[dataType]:
                 continue
@@ -107,6 +114,8 @@ def athenaMPOutputHandler(athenaMPFileReport, athenaMPWorkerTopDir, dataDictiona
             fileNameList = []
             for entry in MPdirWalk:
                 if "evt_count" in entry[0]:
+                    continue
+                if "range_scatterer" in entry[0]:
                     continue
                 # N.B. AthenaMP may have made the output name unique for us, so 
                 # we need to treat the original name as a prefix
@@ -123,7 +132,7 @@ def athenaMPOutputHandler(athenaMPFileReport, athenaMPWorkerTopDir, dataDictiona
                 pass
             elif len(fileNameList) != athenaMPworkers:
                 raise trfExceptions.TransformExecutionException(trfExit.nameToCode("TRF_OUTPUT_FILE_ERROR"), "Found {0} output files for {1}, expected {2} (found: {3})".format(len(fileNameList), dataType, athenaMPworkers, fileNameList))
-            
+
             # Found expected number of files - good!
             athenaMPoutputsLinkAndUpdate(fileNameList, fileArg) 
 
@@ -147,12 +156,19 @@ def athenaMPoutputsLinkAndUpdate(newFullFilenames, fileArg):
             
     for linkname, fname in zip(linkedNameList, newFullFilenames):
         if linkname:
-            try:
-                if path.lexists(linkname):
-                    os.unlink(linkname)
-                os.symlink(fname, linkname)
-            except OSError, e:  
-                raise trfExceptions.TransformExecutionException(trfExit.nameToCode("TRF_OUTPUT_FILE_ERROR"), "Failed to link {0} to {1}: {2}".format(fname, linkname, e))
+            if len(newFullFilenames) == 1:
+                try:
+                    os.rename(fname,fileArg.originalName)
+                    newFilenameValue[0]=fileArg.originalName
+                except OSError, e:
+                    raise trfExceptions.TransformExecutionException(trfExit.nameToCode("TRF_OUTPUT_FILE_ERROR"), "Failed to move {0} to {1}: {2}".format(fname, linkname, e))
+            else:
+                 try:
+                     if path.lexists(linkname):
+                         os.unlink(linkname)
+                     os.symlink(fname, linkname)
+                 except OSError, e:  
+                     raise trfExceptions.TransformExecutionException(trfExit.nameToCode("TRF_OUTPUT_FILE_ERROR"), "Failed to link {0} to {1}: {2}".format(fname, linkname, e))
 
     fileArg.multipleOK = True
     fileArg.value = newFilenameValue
